+++ src/dactyl_keyboard/cad/tweak.clj
@@ -100,17 +100,22 @@
   type of anchor and secondary parameters about the target detail upon it.
   By default, use the most specific dimensions available for the post,
   defaulting to a post for key-cluster webbing."
-  [getopt {:keys [anchoring size] :as node}]
+  [getopt {:keys [anchoring intrinsic-rotation size]
+           :or {intrinsic-rotation [0 0 0]}
+           :as node}]
   {:pre [(spec/valid? ::valid/tweak-leaf node)]}
   (let [{:keys [anchor side segment offset]} anchoring
         {::anch/keys [type primary] :as resolved} (resolve-anchor getopt anchor)
-        default (fn [& shapes] (if size (apply model/cube size)
-                                        (first (filter some? shapes))))]
+        shape (fn [& shapes]
+                "Prefer the tweak’s overriding size to any default shapes."
+                (maybe/rotate intrinsic-rotation
+                  (if size (apply model/cube size)
+                           (first (filter some? shapes)))))]
     (case type
       ::anch/key-mount
         [false
          (let [{:keys [cluster coordinates]} resolved]
-           (web-post getopt cluster coordinates (or side ::anch/any)))]
+           (shape (web-post getopt cluster coordinates (or side ::anch/any))))]
       ::anch/central-gabel
         [true
          (central/tweak-post getopt anchor)]
@@ -120,32 +125,32 @@
       ::anch/rear-housing
         [true
          (place/rhousing-place getopt (::anch/layer resolved) side segment offset
-           (default misc/nodule))]
+           (shape misc/nodule))]
       ::anch/mcu-grip
         [false
-         (default (apply model/cube (getopt :mcu :support :grip :size)))]
+         (shape (apply model/cube (getopt :mcu :support :grip :size)))]
       ;; If a side of the MCU plate is specifed, put a nodule there,
       ;; else use the entire base of the plate.
       ::anch/mcu-lock-plate
         [false
          (if side
-           (default misc/nodule)
-           (default (mcu/lock-plate-base getopt false)))]
+           (shape misc/nodule)
+           (shape (mcu/lock-plate-base getopt false)))]
       ::anch/wr-block
         [true
          (let [{:keys [mount-index block-key]} resolved]
            (place/wrist-block-place getopt mount-index block-key
                                     side segment offset
              (if (or side segment offset)
-               (default misc/nodule)
-               (default (wrist/block-model getopt mount-index block-key)))))]
+               (shape misc/nodule)
+               (shape (wrist/block-model getopt mount-index block-key)))))]
       ::anch/wr-nut
         ;; Ignore side and segment as inapplicable to a nut.
         [true
          (let [{:keys [mount-index block-key fastener-index]} resolved]
            (place/wrist-nut-place getopt mount-index block-key fastener-index
                                   offset
-             (default (wrist/nut getopt mount-index block-key fastener-index))))]
+             (shape (wrist/nut getopt mount-index block-key fastener-index))))]
       ::anch/port-hole
         [true
          (place/port-place getopt anchor
@@ -153,21 +158,21 @@
              (maybe/translate (place/port-hole-offset getopt anchoring)
                ;; Use a nodule by default for tenting the ceiling slightly,
                ;; as would be useful for DFM.
-               (default misc/nodule))
-             (default (auxf/port-hole-base getopt anchor))))]
+               (shape misc/nodule))
+             (shape (auxf/port-hole-base getopt anchor))))]
       ::anch/port-holder
         [true
          (place/port-place getopt primary
            (if (or side segment offset)
              (maybe/translate (place/port-holder-offset getopt
                                 (assoc anchoring :anchor primary))
-               (default (auxf/port-tweak-post getopt primary)))
-             (default (auxf/port-holder getopt primary))))]
+               (shape (auxf/port-tweak-post getopt primary)))
+             (shape (auxf/port-holder getopt primary))))]
       ::anch/secondary
         [false
          (let [{:keys [size]} (getopt :secondaries anchor)]
-           (default (when size (apply model/cube size)) misc/nodule))]
-      [false (default misc/nodule)])))
+           (shape (when size (apply model/cube size)) misc/nodule))]
+      [false (shape misc/nodule)])))
 
 (defn- leaf-blade-3d
   "One model at one vertical segment of one feature."