+++ CHANGELOG.md
@@ -163,6 +163,7 @@ version 0.2.0, thus covering only a fraction of the project’s history.
     - The ability to target ports, including port holders, and the plate of an
       MCU lock, including a set of vertical segments. The rear housing has also
       gotten vertical segments.
+    - Intrinsic rotation.
 - Support for specifying sizes with less than 3-tuples.
 - Support for specifying angles with mathematical formulae using π.
 - A `size` property of `secondaries`, for terser `tweaks`.